+++ src/main/java/model/http/type/HttpMethod.java
@@ -1,4 +1,4 @@
-package model.http.request;
+package model.http.type;
 
 import exception.http.method.NotSupportHttpMethodException;
