+++ src/main/java/webserver/RequestHandler.java
@@ -2,9 +2,9 @@ package webserver;
 
 import model.User;
 import model.http.request.Cookie;
-import model.http.request.HttpMethod;
+import model.http.type.HttpMethod;
 import model.http.request.HttpRequest;
-import model.http.request.MediaType;
+import model.http.type.MediaType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import service.UserService;
@@ -27,7 +27,7 @@ public class RequestHandler extends Thread {
 
   private static final Logger log = LoggerFactory.getLogger(RequestHandler.class);
 
-  private Socket connection;
+  private final Socket connection;
 
   private final UserService userService;
